--story=114904715 OP接口：续期体验时长，用于流水线上执行

+++ src/backend/ci/ext/tencent/remotedev/api-remotedev-tencent/src/main/kotlin/com/tencent/devops/remotedev/api/op/OpRemoteDevResource.kt
@@ -135,7 +135,7 @@ interface OpRemoteDevResource {
         userId: String,
         @ApiParam(value = "续期时长", required = true)
         @QueryParam("renewalTime")
-        renewalTime: Long
+        renewalTime: Int
     ): Result<Boolean>
 
     @ApiOperation("获取用户设置")
